refactor(server): re-cut the padi hosting arm into padi/ (L27, pure motion)
The padi hosting arm — the local + remote binders, their convergence
declaration, session shape and link projection, plus the pool satellites
(reServeEviction, supervisorClaim, daemonInventory) — co-varies on every
remote/switch phase yet sat flat beside the thin HTTP shell. Home it under
packages/server/src/padi/ so the directory names the volatility the flat list
hid (W4 ledger L27). The serving shell (index/router/surface/iframePreviewRoute/
routeErrors) and the true leaves (hostname/tls/state/log/memorySampler/
pwaIdentity) stay top-level — no misc/util bucket.

Strictly moves-only: the sole content edits are the mechanical consequences of
the move — index.ts's import paths, the moved files' ../log.ts reach-up,
padiBinding.test's ../router.ts, and the seal's WEB_SHELL_FILES allowlist
repathed to padi/*. No logic changes; the arm stays inside the package seal
(it re-serves padi's surface, runs no terminal domain). Tests move with their
subjects; bootOrdering/exportTranscript/shell-leaf tests stay top.

hostPool/perHostKoluCells named in the entry don't exist as files (the pool is
buildRemotePool from @kolu/surface-remote, instantiated inline in index.ts) —
verified against the real tree; the cut follows the code, not the sketch.
