+++ packages/server/src/padi/daemonInventory.test.ts
