+++ packages/server/src/padi/padiBinding.ts
@@ -78,7 +78,7 @@ import {
   measureClockOffset,
   type Session,
 } from "@kolu/surface-remote";
-import { log } from "./log.ts";
+import { log } from "../log.ts";
 // padi's convergence declaration into the shared daemon-convergence kit — the
 // contract-skew POLICY, the FROZEN-control-core probe, and the drain plumbing the
 // probe and the "restart" verb share. Carved out of this file in W4 ledger L6: it
